+++ lib/vmobject.dart
@@ -1,3 +1,4 @@
 export 'package:obsobject/obsobject.dart';
+export 'src/observerFormField.dart';
 export 'src/observerWidget.dart';
 export 'src/viewModel.dart';